hw3/p2_requirements.txt
gdown
imageio==2.21.3
matplotlib==3.6.1
numpy==1.23.4
Pillow==9.2.0
scipy==1.9.1
opencv-python==4.6.0.66
loralib==0.1.2
pycocotools==2.0.5
timm==0.9.10
torch==1.12.1
pandas==1.5.1
tqdm
git+https://github.com/openai/CLIP.git
git+https://github.com/bckim92/language-evaluation.git